+++ cmd/platone/usage.go
@@ -221,10 +221,6 @@ var AppHelpFlagGroups = []flagGroup{
 		},
 	},
 	{
-		Name:  "WHISPER (EXPERIMENTAL)",
-		Flags: whisperFlags,
-	},
-	{
 		Name: "DEPRECATED",
 		Flags: []cli.Flag{
 			utils.MinerLegacyThreadsFlag,